Move the examples out of the way of the editor - avoid clashes on rendering
