+++ src/Links.tsx
@@ -1,42 +1,11 @@
 import * as React from "react";
 import { Component } from "react";
 import { Box, Divider } from "rendition";
-import ExampleLinks from "./ExampleLinks";
 
-import ky from "ky";
-
-function importAll(r: any) {
-  return r.keys().map(r);
-}
-
-const exampleFilePaths = importAll(require.context("../examples", true, /\.(yaml)$/));
-
-interface IHash {
-  [key: string]: string
-}
-
-let allExamples: IHash = {};
-
-interface LinksState {
-  allExamples: IHash
-}
-
-class Links extends Component<any, LinksState> {
+class Links extends Component<any, any> {
 
   constructor(props: any) {
     super(props);
-    this.state = { allExamples: {} };
-  }
-
-  componentDidMount(): void {
-    exampleFilePaths.map((path: string) => {
-      ky.get(path).text().then((content) => {
-        const name = path.replace("examples/", "");
-        allExamples[name] = content;
-      }).then(() => {
-        this.setState({ allExamples: allExamples });
-      });
-    });
   }
 
   render(): React.ReactNode {
@@ -79,7 +48,6 @@ class Links extends Component<any, LinksState> {
           Known UI issues
         </a>
         <Divider/>
-        <ExampleLinks examples={this.state.allExamples}/>
       </Box>
     );
   }
